Introduce a way to pass config down into the plugins
Use this to implement a way to keep the current stash when doing
a subrequest or visit. Docs and tests to follow.

This is also pretty gross, it'd be nicer at this point to make the
plugins instances which are built (and their config merged) on first
use, rather than merging and passing the config down into the class
method every time.